Refactor Beta page
