+++ modules/Feed.module.jsx
@@ -1,5 +1,5 @@
 import AppWrapper from "./AppWrapper";
-import Beta from "../Beta";
+import Beta from "./Beta";
 import Feed from "../components/Feed.jsx";
 import Loading from "../components/Loading";
